+++ src/biodwh2-core/src/main/java/de/unibi/agbi/biodwh2/core/DataSource.java
@@ -51,6 +51,10 @@ public abstract class DataSource {
         return null;
     }
 
+    public String[] getDependencies() {
+        return null;
+    }
+
     public abstract DevelopmentState getDevelopmentState();
 
     protected abstract Updater<? extends DataSource> getUpdater();